+++ server/src/internalClusterTest/java/org/opensearch/action/admin/cluster/shards/TransportCatShardsActionIT.java
@@ -15,12 +15,17 @@ import org.opensearch.cluster.routing.ShardRouting;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.core.action.ActionListener;
 import org.opensearch.core.common.Strings;
+import org.opensearch.core.tasks.TaskCancelledException;
+import org.opensearch.test.InternalTestCluster;
 import org.opensearch.test.OpenSearchIntegTestCase;
 
+import java.io.IOException;
 import java.util.List;
 
 import static org.opensearch.cluster.routing.UnassignedInfo.INDEX_DELAYED_NODE_LEFT_TIMEOUT_SETTING;
+import static org.opensearch.common.unit.TimeValue.timeValueMillis;
 import static org.opensearch.search.SearchService.NO_TIMEOUT;
+import static org.junit.Assert.fail;
 
 @OpenSearchIntegTestCase.ClusterScope(numDataNodes = 0, scope = OpenSearchIntegTestCase.Scope.TEST)
 public class TransportCatShardsActionIT extends OpenSearchIntegTestCase {
@@ -54,6 +59,58 @@ public class TransportCatShardsActionIT extends OpenSearchIntegTestCase {
 
             @Override
             public void onFailure(Exception e) {}
+        });
+    }
+
+    public void testCatShardsWithTimeoutException() throws IOException {
+        List<String> masterNodes = internalCluster().startClusterManagerOnlyNodes(1);
+        List<String> nodes = internalCluster().startDataOnlyNodes(3);
+        createIndex(
+            "test",
+            Settings.builder()
+                .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 1)
+                .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 2)
+                .put(INDEX_DELAYED_NODE_LEFT_TIMEOUT_SETTING.getKey(), "60m")
+                .build()
+        );
+        ensureGreen("test");
+
+        Settings clusterManagerDataPathSettings = internalCluster().dataPathSettings(masterNodes.get(0));
+        // Dropping master node to delay in cluster state call.
+        internalCluster().stopRandomNode(InternalTestCluster.nameFilter(masterNodes.get(0)));
+
+        new Thread(() -> {
+            try {
+                // Ensures the cancellation timeout expires.
+                Thread.sleep(1000);
+                // Starting master node to proceed in cluster state call.
+                internalCluster().startClusterManagerOnlyNode(
+                    Settings.builder().put("node.name", masterNodes.get(0)).put(clusterManagerDataPathSettings).build()
+                );
+            } catch (InterruptedException e) {
+                throw new RuntimeException(e);
+            }
+        }).start();
+
+        final CatShardsRequest shardsRequest = new CatShardsRequest();
+        shardsRequest.setCancelAfterTimeInterval(timeValueMillis(1000));
+        shardsRequest.setIndices(Strings.EMPTY_ARRAY);
+        client().execute(CatShardsAction.INSTANCE, shardsRequest, new ActionListener<CatShardsResponse>() {
+            @Override
+            public void onResponse(CatShardsResponse catShardsResponse) {
+                // onResponse should not be called.
+                fail();
+            }
+
+            @Override
+            public void onFailure(Exception e) {
+                boolean timeoutException = (e.getClass() == TaskCancelledException.class);
+                if (e.getCause() != null) {
+                    timeoutException = timeoutException
+                        || (e.getCause().getMessage().equals("The parent task was cancelled, shouldn't start any child tasks"));
+                }
+                assertTrue(timeoutException);
+            }
         });
     }