+++ server/src/main/java/org/opensearch/action/admin/cluster/shards/TransportCatShardsAction.java
@@ -23,6 +23,13 @@ import org.opensearch.tasks.CancellableTask;
 import org.opensearch.tasks.Task;
 import org.opensearch.transport.TransportService;
 
+import java.util.concurrent.CountDownLatch;
+
+/**
+ * Perform cat shards action
+ *
+ * @opensearch.internal
+ */
 public class TransportCatShardsAction extends HandledTransportAction<CatShardsRequest, CatShardsResponse> {
 
     private final TransportService transportService;
@@ -60,32 +67,46 @@ public class TransportCatShardsAction extends HandledTransportAction<CatShardsRe
         }
         CatShardsResponse catShardsResponse = new CatShardsResponse();
 
+        CountDownLatch waitForClusterState = new CountDownLatch(1);
+
         client.admin().cluster().state(clusterStateRequest, new ActionListener<ClusterStateResponse>() {
             @Override
             public void onResponse(ClusterStateResponse clusterStateResponse) {
-                IndicesStatsRequest indicesStatsRequest = new IndicesStatsRequest();
-                indicesStatsRequest.all();
-                indicesStatsRequest.indices(shardsRequest.getIndices());
                 catShardsResponse.setClusterStateResponse(clusterStateResponse);
-                indicesStatsRequest.setParentTask(client.getLocalNodeId(), task.getId());
-                client.admin().indices().stats(indicesStatsRequest, new ActionListener<IndicesStatsResponse>() {
-                    @Override
-                    public void onResponse(IndicesStatsResponse indicesStatsResponse) {
-                        catShardsResponse.setIndicesStatsResponse(indicesStatsResponse);
-                        listener.onResponse(catShardsResponse);
-                    }
-
-                    @Override
-                    public void onFailure(Exception e) {
-                        listener.onFailure(e);
-                    }
-                });
+                waitForClusterState.countDown();
             }
 
             @Override
             public void onFailure(Exception e) {
+                waitForClusterState.countDown();
                 listener.onFailure(e);
             }
         });
+
+        try {
+            // Ensures that cluster state transport action completed.
+            waitForClusterState.await();
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
+        }
+        if (catShardsResponse.getClusterStateResponse() != null) {
+            IndicesStatsRequest indicesStatsRequest = new IndicesStatsRequest();
+            indicesStatsRequest.all();
+            indicesStatsRequest.indices(shardsRequest.getIndices());
+            indicesStatsRequest.setParentTask(client.getLocalNodeId(), task.getId());
+            client.admin().indices().stats(indicesStatsRequest, new ActionListener<IndicesStatsResponse>() {
+                @Override
+                public void onResponse(IndicesStatsResponse indicesStatsResponse) {
+                    catShardsResponse.setIndicesStatsResponse(indicesStatsResponse);
+                    listener.onResponse(catShardsResponse);
+                }
+
+                @Override
+                public void onFailure(Exception e) {
+                    listener.onFailure(e);
+                }
+            });
+        }
+
     }
 }